consolidating migrations
