+++ migration/src/lib.rs
@@ -1,16 +1,12 @@
 pub use sea_orm_migration::prelude::*;
 
 mod m20220101_000001_create_table;
-mod m20230605_200842_create_memory_table;
 
 pub struct Migrator;
 
 #[async_trait::async_trait]
 impl MigratorTrait for Migrator {
     fn migrations() -> Vec<Box<dyn MigrationTrait>> {
-        vec![
-            Box::new(m20220101_000001_create_table::Migration),
-            Box::new(m20230605_200842_create_memory_table::Migration),
-        ]
+        vec![Box::new(m20220101_000001_create_table::Migration)]
     }
 }
